backend/requirements.txt
Flask==3.0.3
Flask-Cors==3.0.7
pymongo==4.12.1
requests==2.32.3
python-dotenv==1.1.1
joblib==1.4.2
lightgbm==4.6.0
